requirements.txt
langchain-community
google-generativeai
sentence-transformers
uvicorn==0.19.0
fastapi==0.111.0
pydantic==2.9.2
pytest==6.2.5
python-dotenv
